refactor: Move DI config into config classes
Moving config into classes allows us to work with the aura/di lifecycle properly
and configure everything before the container is locked.
