+++ app/Providers/AppServiceProvider.php
@@ -26,8 +26,6 @@ class AppServiceProvider extends ServiceProvider
         $this->registerFilesystem();
         $this->registerMailer();
 
-        $this->configureAuraDI();
-
         $this->registerDataSources();
 
         $this->setupMultisiteIlluminateDB();
@@ -102,98 +100,7 @@ class AppServiceProvider extends ServiceProvider
         $this->app->register(\Ushahidi\App\DataSource\DataSourceServiceProvider::class);
     }
 
-    protected function configureAuraDI()
-    {
-        $di = service();
-
-        $this->configureAuraServices($di);
-        $this->injectAuraConfig($di);
-    }
-
-    protected function configureAuraServices(\Aura\Di\Container $di)
-    {
-        // Configure mailer
-        $di->set('tool.mailer', $di->lazyNew('Ushahidi\App\Tools\LumenMailer', [
-            'mailer' => app('mailer'),
-            'siteConfig' => $di->lazyGet('site.config'),
-            'clientUrl' => $di->lazyGet('clienturl')
-        ]));
-
-        // Configure filesystem
-        // The Ushahidi filesystem adapter returns a flysystem adapter for a given
-        // cdn type based on the provided configuration
-        $di->set('tool.filesystem', function () {
-            // Get the underlying League\Flysystem\Filesystem instance
-            return $this->app->make('filesystem')->disk()->getDriver();
-        });
-
-        // Setup user session service
-        $di->set('session', $di->lazyNew(\Ushahidi\App\Tools\LumenSession::class, [
-            'userRepo' => $di->lazyGet('repository.user')
-        ]));
-
-        // Multisite db
-        $di->set('kohana.db.multisite', function () use ($di) {
-            $config = config('ohanzee-db');
-
-            return \Ohanzee\Database::instance('multisite', $config['multisite']);
-        });
-
-        // Deployment db
-        $di->set('kohana.db', function () use ($di) {
-            return \Ohanzee\Database::instance('deployment', $this->getDbConfig($di));
-        });
-
-        // Configure dispatcher
-        $di->setters[\Ushahidi\Core\Traits\Events\DispatchesEvents::class]['setDispatcher']
-            = $this->app->make('events');
-    }
-
-    protected function injectAuraConfig(\Aura\Di\Container $di)
-    {
-        // CDN Config settings
-        $di->set('cdn.config', function () use ($di) {
-            return config('cdn');
-        });
-
-        // Ratelimiter config settings
-        $di->set('ratelimiter.config', function () use ($di) {
-            return config('ratelimiter');
-        });
-
-        // Multisite db
-        // Move multisite enabled check to class and move to src/App
-        $di->set('site', function () use ($di) {
-            // @todo default to using the current domain
-            $site = '';
-
-            // Is this a multisite install?
-            $multisite = config('multisite.enabled');
-            if ($multisite) {
-                $site = $di->get('multisite')->getSite();
-            }
-
-            return $site;
-        });
-
-        // Move multisite enabled check to class and move to src/App
-        $di->set('tool.uploader.prefix', function () use ($di) {
-            // Is this a multisite install?
-            $multisite = config('multisite.enabled');
-            if ($multisite) {
-                return $di->get('multisite')->getCdnPrefix();
-            }
-
-            return '';
-        });
-
-        // Client Url
-        $di->set('clienturl', function () use ($di) {
-            return $this->getClientUrl($di->get('site.config'), $di->lazyGet('multisite'));
-        });
-    }
-
-    protected function getDbConfig(\Aura\Di\Container $di)
+    protected function getDbConfig()
     {
         // Kohana injection
         // DB config
@@ -203,7 +110,7 @@ class AppServiceProvider extends ServiceProvider
         // Is this a multisite install?
         $multisite = config('multisite.enabled');
         if ($multisite) {
-            $config = $di->get('multisite')->getDbConfig();
+            $config = service('multisite')->getDbConfig();
         }
 
         return $config;
@@ -230,8 +137,7 @@ class AppServiceProvider extends ServiceProvider
 
     protected function setupMultisiteIlluminateDB()
     {
-        $di = service();
-        $config = $this->getDbConfig($di);
+        $config = $this->getDbConfig();
 
         $existing = config('database.connections.mysql');
