--- src/Console/Init.php
@@ -1,5 +1,0 @@
-<?php
-
-$di = service();
-
-$di->set('authorizer.console', $di->lazyNew('Ushahidi\Console\Authorizer\ConsoleAuthorizer'));
